remove go publish legacy
